Added image for notes. Save image. Binding note to the folder!

+++ LocationNotes/LocationNotes/NoteController.swift
@@ -16,17 +16,34 @@ class NoteController: UITableViewController {
     @IBOutlet weak var textName: UITextField!
     @IBOutlet weak var textDescription: UITextView!
     
+    @IBOutlet weak var labelFolder: UILabel!
+    @IBOutlet weak var labelFolderName: UILabel!
+    
     override func viewDidLoad() {
         super.viewDidLoad()
 
         textName.text = note?.name
         textDescription.text = note?.textDescription
+        imageView.image = note?.imageActual
+        
+        navigationItem.title = note?.name
      
     }
     
-    override func viewWillDisappear(_ animated: Bool) {
-        
-        if textName.text == "" && textDescription.text == "" {
+    override func viewWillAppear(_ animated: Bool) {
+        if let folder = note?.folder {
+            labelFolderName.text = folder.name
+        }else {
+            labelFolderName.text = "-"
+        }
+    }
+
+    @IBAction func pushDoneAction(_ sender: Any) {
+        saveNote()
+        _ = navigationController?.popViewController(animated: true)
+    }
+    func saveNote() {
+        if textName.text == "" && textDescription.text == "" && imageView.image == nil{
             CoreDataManager.sharedInstance.managetObjectContext.delete(note!)
             CoreDataManager.sharedInstance.saveContext()
             return
@@ -39,7 +56,57 @@ class NoteController: UITableViewController {
         note?.name = textName.text
         note?.textDescription = textDescription.text
         
+        note?.imageActual = imageView.image
+        
         CoreDataManager.sharedInstance.saveContext()
+    }
+    
+    let imagePicker: UIImagePickerController = UIImagePickerController()
+    
+    
+    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        
+        tableView.deselectRow(at: indexPath, animated: true)
+        
+        if indexPath.row == 0 && indexPath.section == 0 {
+            let alertController = UIAlertController(title: "Added image for item", message: "", preferredStyle: UIAlertController.Style.actionSheet)
+            
+            
+            let alertOneCamera = UIAlertAction(title: "Make a photo", style: UIAlertAction.Style.default) { (alert) in
+                print("Send to Photo with camera")
+                self.imagePicker.sourceType = .camera
+                self.imagePicker.delegate = self
+                self.present(self.imagePicker, animated: true, completion: nil)
+            }
+            
+            alertController.addAction(alertOneCamera)
+            
+            let alertTwoPhoto = UIAlertAction(title: "Select from library", style: UIAlertAction.Style.default) { (alert) in
+                //print("Send to library Photo")
+                
+                self.imagePicker.sourceType = .savedPhotosAlbum
+                self.imagePicker.delegate = self
+                self.present(self.imagePicker, animated: true, completion: nil)
+            }
+            alertController.addAction(alertTwoPhoto)
+
+             if self.imageView.image != nil{
+                    let alertTheeDelete = UIAlertAction(title: "Delete", style: UIAlertAction.Style.destructive) { (alert) in
+                        self.imageView.image = nil
+                    }
+                 alertController.addAction(alertTheeDelete)
+            }
+            
+            
+            let alertFourCancel = UIAlertAction(title: "Cancel", style: UIAlertAction.Style.cancel) { (alert) in
+                
+            }
+            
+            alertController.addAction(alertFourCancel)
+            
+            present(alertController, animated: true, completion: nil)
+            
+        }
     }
 
     // MARK: - Table view data source
@@ -76,6 +143,7 @@ class NoteController: UITableViewController {
     // Override to support editing the table view.
     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
         if editingStyle == .delete {
+            let deletedNote
             // Delete the row from the data source
             tableView.deleteRows(at: [indexPath], with: .fade)
         } else if editingStyle == .insert {
@@ -99,14 +167,27 @@ class NoteController: UITableViewController {
     }
     */
 
-    /*
+    
     // MARK: - Navigation
 
     // In a storyboard-based application, you will often want to do a little preparation before navigation
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
+        if segue.identifier == "goToSelectFolder" {
+            (segue.destination as! SelectFolderController).note = note
+        }
     }
-    */
+    
 
+}
+
+extension NoteController: UIImagePickerControllerDelegate, UINavigationControllerDelegate {
+    
+    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
+        imageView.image = info[UIImagePickerController.InfoKey.originalImage] as? UIImage
+        picker.dismiss(animated: true, completion: nil)
+    }
+    
+    func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
+        picker.dismiss(animated: true, completion: nil)
+    }
 }
